[VTK-4071] Customizable query in SubFolderProvider
VTK-4071 requires a subfolder menu that provides all subfolders except the ones with resource type 'internal-organizational-unit-folder'. With this change the a query filter can be added in the bean config as a String.
The resource type 'internal-organizational-unit-folder' is also added with this issue. Thus additions in @collection.xml and @namespaces.xml were also required in the vtk project.