Remove 'requireSerializable' option for EhContentCache.
The Ehcache API for enforcing serializable keys/values is deprecated, and we do
not have a single cache configured with this property set to 'true'. So the
option is removed altogether.

(Just make sure the keys/values *are* in fact serializable if the Ehcache is
configured for off-heap storage.)

Also, increase background refresh interval to 60 seconds for most of our content
caches, which should be more than enough, considering that we typically operate
with 300 seconds time-to-live.