[VTK-4099] No principal null check for Rev retrive
Removed check that would throw exception when principal is null in repository retrive function. Repository retrive will still use principal for acl validation. This was done to make components (person-language-link in this case) not return exception messages for version diff. HEAD version would not return this exception message and it would always be marked as a difference from its revisions. Between revisions the exception message would be visible but not marked as a difference as both
would show it.