[VTK-4164] Make Comment.author a Principal again.
CommentDAO should populate Comment objects with Principal instances for
the author field to be consistent with the rest of our repository DAO
code. This fixes Freemarker view code expecting Comment authors to be
Principals.

Implementation details: Add a MyBatis type handler which is able to map
UID strings in database to USER Principals. This way we can keep using
MyBatis to map Comment instances directly.

Note that populating Principal instances from the DAO layer is not a
good thing, but it's better to be consistent now, before we eventually
have time to refactor the whole DAO Principal mess.