ServiceUrlProvider: best effort when missing RC.
Unfortunately a thread local request context is not always available in
threads calling service URL provider. This is required for Service URL
construction to work. Make it do best effort if request context is
missing, by using default URL map. If that fails as well, then throw an
exception describing the problem.