[VTK-4299] Make data in TokenManager cache serializable.
Since the handlers themselves are not serializable across a distributed
cache, we maintain a local map based on the handler id. 

NB! This assumes the handler identifiers are unique. This map should
probably be injected by Spring so that all possible handlers are know
early. Otherwise a different node in the cluster may not have the
handler mapping already.