New IO utility
Add a new I/O utility, `vtk.util.io.IO` which is meant to replace all
usage of `vtk.util.io.StreamUtil`. It has more features, better API and
better extensibility.

Specifically, it allows progress tracking for all operations [VTK-4425],
and all operations support the general concepts of offset and limit.

It comes with a barrage of tests in `IOTest`, mostly based on the tests
for existing `StreamUtil`, with some more added. The tests are also
examples on usage.