Repo op log: use annotations instead of hard coded mess
Make operation log interceptor look at annotations in RepositoryImpl to
determine of the operation should be logged and which parameter the log
should include.

This makes it safe to refactor Repository operation methods without
having to worry about updating the operation log mess in addition. It's
also a lot more flexible.