[VTK-4444] Run StandardRequestFilter earlier
Spring Boot: Implement javax.servlet.Filter in StandardRequestFilter,
and run it first in filter chain.

This is necessary because spring-session also uses a servlet filter, and
we need to distinguish between http and https requests when setting
session cookies (different names for different protocols).

Also separate out 'stdRequestFilter' bean and run it as usual when not
in spring-boot.