[VTK-4727] Use file system move to store content when possible
Introduces `ContentInputSource` interface for use when storing to repository.
This interface allows setting a file source and if the source file can be
deleted after storing.

If the source is a file which client code allows the repository to delete after
storing (typically temporary files outside of repository), then the file will be
moved atomically into the content store. If such move operation fails, it falls back
to regular stream copying.

Also adds `PosixFileSystemContentStore`, which can be used when custom file and
directory permissions are required on the content store data files. (This is not
configured by default.)