[VTK-4816] Ignore client disconnect
Client disconnects currently introduces large amounts of error log clutter which
we can do nothing about. It also causes request to be logged as "500 internal
server error", which is mostly false.

This commit catches java.io.EOFException in VTKServlet, which seems to be a good
candidate for handling these situations with a lot less noise. Requests where
client disconnects will be logged as "200 OK" in access log, if server processed
it successfully, but a VTKServlet will spit out a small WARNing.