[VTK-4805] Fix up misunderstood usage of Cloneable and Object#clone()
We implemented overridden versions of Object#clone without taking advantage of
the built-in magic for this method. Always call super.clone(), eventually
leading to Object#clone in case of hierarchies, which does a complete shallow
field copy of whatever the runtime class of the object is. Then do steps for
deeper cloning where necessary. Now we can remove the entire clone override in
ResourceImpl, since only the property map needs deeper cloning, which is handled
in the super class, and in general the clone methods become much smaller.

In addition to this, reduce need for deeper cloning by making some classes
immutable:
- BufferedBinaryValue, which also makes Value really immutable (perhaps except
  for Principal, but we ignore that until principal cleanup happens)
- Lock (removed LockImpl, made Lock a real immutable class instead)