[VTK-4805] Remove ChildDelegatingMessourceSource
Add basenames to a single standard, but slightly extended
ResourceBundleMessageSource instead. (The extension is a simple locale
string-representation mapping facility, which we likely need to avoid breakage.)

The benefits are:
- likely much improved message lookup performance, since most things will be
cached in a common source (ChildDelegatingMessageSource iterated all sources on
context for every single lookup, every time..).
- Less XML when using the "method invoking bean builder" config pattern, which
adds basenames to the common source, instead of adding a new
ResourceBundleMessageSource in the context.
- Less magic. To make your messages available in global source (for Freemarker,
etc), you must simply register your resource bundle basename(s) in the global
source.