[VTK-4837] Do auto-reindex when incompatible changes are detected
Write an application level version number when indexes are created. Check this
number on init and possibly do automatic reindexing if the number has been
bumped in the code base.

All logic related to automatic reindexing at initialization time has been
simplified and moved into IndexOperationManagerImpl. (ReindexAtStartupBean is
gone.)

Also, for consistency checks, don't be rude and load a potentially huge set of
strings into memory (database URIs). Instead, use OrderedIndexSet, which is a
utility that can be used as an on-disk set of string values (impl based on a
temporary Lucene index).