[VTK-4837] Add DocValues for single value numeric types
Add DocValue fields for all single value numeric types. Lucene will
automatically use these for sorting if available.

When available and used for sorting, Lucene does not need to uninvert the
regular indexed field and cache doc->value mappings in memory. This should shift
memory usage off the heap and over to OS page cache (memory mapped index files).
(It also makes the on disk index bigger and indexing more costly, but as usual
we optimize for fast reads and queries primarily, and reducing heap pressure is
a good thing as well.)