[VTK-4878] Implement support for update waiting for index searches
Added an API in Search for waiting for index to become updated before the search
is executed, which can block the searching thread.

The API accepts a timestamp for a required recency of the search results, which
will typically be "Instant.now()", but can be set to anything. In addtion, a
timeout is required so we never wait forever. (Though it permits huge timeouts
which can practically be forever..)

Overview of algorithm for searches with a wait-spec set:
1. Check database change log for any events older than provided timestamp.
2. If present, then use a new API in IncrementalUpdater, which allows a thread
to wait for the next batch of updates to complete (with a timeout).
3. If waited for batch and deadline (client timeout) has not been reached, check
database changelog again and see if oldest is now more recent than provided
timestamp. If deadline reached, then return and execute search we an estimated
recency set to the oldest changelog event the last time we checked.

Points 1-3 are executed in a loop while current time is within deadline.

An estimated (and Optional) recency date is set on the ResultSet of the search,
which can be used by client code to determine if the search was executed on an
index up to date as of the timestamp specified or not (in case timeout deadline
was reached).

Other changes:
* IncrementalUpdater now does its own update thread management and no longer
relies on "SimpleMethodInvokingTriggerBean".
* ChangeLogDao API updates to allow event timestamp filtering.
* Probably some more things..