Remove ResourceTypeDefinition#getQName()
Qualified name does not make sense for resource type names, since they must
always be unique with a repository instance. So they don't need to be
namespace-qualified.

Document intention behind setting a namespace on a resource type definition,
which is mainly for enforcing a namespace on the property definitions it has.