Fix serious bug in ArrayStackCache when used without default factory
Silly bug and poor unit testing.. Lucky for us, most use cases for this cache is
with a default factory set.

- Fix #getInstance() would always return null if no default factory was
  configured at creation time, even though the cache had available objects in
  it.

- Make unit test cover the case of no default factory better.

May be related to VTK-5043, since VortexFastCGIServlet does not use a default
factory for the keepalive socket cache. So no sockets would ever be reused,
since they just sat idling away in the cache. They were not closed in a timely
manner either, as a result.