[VTK-4496] Purge trashed resources within individual transactions
* Periodic maintenance job which purges trash objects uses a separate transaction
for each (soon to be non-) recoverable resource. This ensure that failure to
process overdue and orphan lists does not cause content store deletions which
cannot be rolled back.

Also, change DAO API for recoverable resources so that only one can be deleted
permanently per call, in style with other related API calls.

* Use "<tx:annotation-driven ..>" for setup of Spring transaction management
with @Transactional annotations.

This restores proper handling of transactional attributes like "readOnly =
true", which now actually have an effect on the surrounding tx (before this, we
only got all defaults, which mostly works, but still..).

This also removes the dubious "no-rollback-for"-settings in previous config,
which was more likely just a band-aid on unintentional nesting of transactions
due to interceptor mechanics confusion, where inner transaction failure could
cause "UnexpectedRollbackException"-style issues when outer, which thought
everything was dandy, tried to commit. We now explicitly only AOP concrete
target classes and not the Repository interface, which has many actual
implementations.