[VTK-5113] Add "Most recently created" report
Since we (likely) want to display columns "Created" and "Created by" instead of
"Modified" and "Modified by" for this report, a new Freemarker template was
added, which is mostly a copy of `document-reporter.ftl`. (Comment added in
heading about refactoring out common parts, but I found no simple way to
generalize with regard to flexible column display types.)

Also slightly refactor `LastModifiedReporter` by renaming it to something more
generic, since it can be made to sort on things other than last modified, and it
is also the backend of the "last-created" report.