[VTK-5091] Make RequestContext no longer a ThreadLocal
Bind the request context to the servlet request in
RequestContextInitializer. The new way of obtaining the context is by
passing the request to RequestContext.getRequestContext().

This has implications for assertions, which are now split in two
hierarchies, and are no longer related by a common superclass:

vtk.web.service.WebAssertion (with access to the request)
vtk.repository.resourcetype.RepositoryAssertion (only performs
matching on resource and principals). For service matching and URL
construction, only WebAssertions are now applicable.

Some other refactorings were also necessary, primarily related to
(system index) query parsing and searching, which in some cases were
written to assume that a request context would be available in a
thread local.