[VTK-5192] Implement new resource lock type SHARED_ACL_WRITE
The Lock class gained a type field, with the classic type named EXCLUSIVE and
the new type called SHARED_ACL_WRITE.

The type of lock is selected when Repository.lock() on a resource is called. All
current code create locks with the EXCLUSIVE type, so there should be no change
in behaviour.

The following applies for locks of the new type: All users which have
write-access, through ACL, to a resource locked with a lock of type
SHARED_ACL_WRITE are authorized to write to the resource, provided that a valid
lock token is also provided in the repository operation. This means that a
single lock token can be shared by several users. Client code must manage such
lock tokens itself, Vortex only validates the incoming tokens against the token
stored on a particular lock.

It is not possible for a regular user to remove a lock of type SHARED_ACL_WRITE
without a valid token, even if the very same user is also the lock owner and has
administrative privileges on a resource. Root role is allowed to remove any type
of lock present.

To support this, a Repository API change was necessary, and all write-operations
now take an additional "String lockToken" parameter, as well as the regular
token. This parameter is null for all cases where the code only operates with
regular locks of type EXCLUSIVE, but should be used for locks of the new type.

For WebDAV, any locks of the new type appear just as regular locks, so it should
be transparent.