VShell: implement "repo id"  in Groovy
Works as a kick starter sample, notice the following differences: The bean
attribute "script-interfaces" is not supported by Spring for Groovy scripts,
instead one must declare the implemented interfaces in the Groovy source code
itself. The Groovy code also needs to declare types that are compatible with the
implemented interfaces. (So, in summary, coercion from loosely typed script
methods to formal interface is slightly less flexible than for BeanShell, but
code is clearer and allows better IDE integration with existing Java code.)