[VTK-5206] Use Groovy compiled for modern JVMs with INvokeDYnamic support
Default Groovy-jars are compatible with old JVMs and do not take advantage of
invokedynamic. This is useless for us, since we are already at Java8+, so use a
set of Groovy-jars that perform better. Add Maven classifier "indy" to all
Groovy dependencies to make it happen. Read about that here:
http://groovy-lang.org/indy.html

Also, parameterize Groovy version in POM.