VShell: disable script refresh check by default
Automatic script refresh causes Spring to create a Proxy around all
"lang:groovy" beans with this attribute set. This is unnecessary, since all
Groovy VCommand beans are actual instances of vtk.shell.vshell.VCommand, and the
interface does not need to be synthesized on top of scripts. Also, the constant
refresh-checking is not necessary for classpath resources in production.
Therefore, set to negative value to disable the functionality. (Unfortunately,
the value can not be parameterized due to "lang:groovy" XML schema restrictions
on the numeric format of this attribute.)

For development, it's easy to adjust refresh check values temporarily anyway.