[VTK-5277] Fix unsafe use of QueryParser in ParsedQueryBuilder
ParsedQueryBuilder and subclasses shared QueryParser instances between multiple
threads, which the stateful JavaCC parser implementation does not support. For
every parse, the factory must be used to create a new parser instance.

This is likely the cause of observed strange error message:

  Failed to parse query "type IN structured-event": Encountered " "IN" "IN "" at line 1

in issue VTK-5277.

The error was introduced in a rather recent commit, which explains why we
haven't seen this issue before.