[VTK-5308] Upgrade to Lucene 7.1.0
This commit upgrades the Vortex repository search backend to use Lucene 7. Many
changes were necssary due to complete removal of search filters in the new
version. We used such filters extensively before, to remove any unnecessary
score computations. With Lucene 7, scores are easier to disable for regular
queries, and filtering is built into the the queries themselves.

* Our custom filter implementations have been removed. One important usage of
filters for us was to invert the matching doc set for any query. Now inversion
is handled by building suitable Lucene boolean query constructs around the query
to invert.

Since Lucene will not match any document for pure negative boolean queries, the
query building process rewrites queries as necessary to provide the "inverse
set" logic to negation (instead of prohibition, which is how Lucene treats it.)

* Field and document mapping code has been reworked for the new version. Numeric
fields are now based on new "PointValues" field type in Lucene, which is not
indexed in regular inversed index structure, but rather as a structure optimized
for efficient numeric queries (k-d tree). Building these queries is now handled
by the field mapping code (and used by the various QueryBuilder impls).

https://www.elastic.co/blog/lucene-points-6.0

* Iteration API impl in LuceneSearcher has been massively simplified as a result
of moving to Lucene 7. All custom segment walking and sorting code is gone, and
we instead use the Lucene "searchAfter(..., scoreDoc)" and HitCollector APIs to
provide the same thing.

Sorted iteration is in essence implemented as efficient "deep paging" in the
search results and no longer has any limitations on the sorting criteria.
Unsorted iteration is simply implemented using a HitCollector that directly
invokes user callback for each match.

* Our custom query and filter caching code has been removed completely. Lucene
now has an internal query cache, which by default is based on sensible defaults
and LRU usage tracking semantics. These things are adjustable of course, but we
currently provide no option to configure it from Spring.

* We no longer use a custom codec which disables compression of stored fields.
The performance impact of this is unknown at this time, but Lucene has been
optimizied in this area as well, selecting different compression codecs based on
data size, so hopefully it will not be an issue at all.

* Robustified code which automatically detects incompatible index and initiates
reindexing. Handle case with Major Lucene version incompatibility.

* Exists queires have become more efficient by using a different strategy than
before: all exists queries are based on term query in a single meta field, which
has as values the names of all other fields in a document. As an added bonus,
exists queries will now work for JSON properties without specifying complex
value attribute (without "..@..") as well.