Delete seemingly unused template file with ref to non-existing component
The template file references component "${document:attributes}" which seems to
have been renamed to "${document:element-attributes}", and so the conclusion is
that this file is no longer in use. There are no direct references to the file
in any source code for VTK, Vortex or vortex-provision. Let's see what happens.