Eliminate HttpUtil#getStatusMessage()
Spring HttpStatus class is more than comprehensive enough to cover all our http
status code needs, including WebDAV, so get rid of our own maintenance of this
in HttpUtil.

This is a pure refactoring change with no modifications to behaviour, except
that it fixes our erroneous mapping of code 418 to "Unprocessable Entity", which
is actually code 422. 418 is reserved for IOT teapots, when a request for coffee
is attempted.