[VTK-5347] Add SamlLoginEvent, SamlLogoutEvent, remove LoginListener
* Remove SAML LoginListener and replace with Spring application events
SamlLoginEvent and SamlLogoutEvent. SamlLoginEvent is published at the same
point that LoginListener instances where invoked before.

For logout, we publish an event for each session (each hostname) that is
SAML-aware, by publishing both when a Vortex host initiates logout process, and
when receiving logout messages as SAMLRequests.

* Cookie monster: replace countless implementations of the common
"getCookieByName"-logic sprinkled all over the code base, with a single
Optional-based implementation in HttpUtil.