[VTK-5387] Implement simple admin template component placeholders
Invocation of the "manage:placeholder"-component is sprinkled at desired
locations in admin templates, with suitable placement identifiers. The component
looks into a registry to determine the components which should be invoked, if
any, for the placement identifier, then invokes each of components in order and
gathers the output. The final output is the output of all invoked components.

Components can then be added to the registry to customize/add to selected parts
of VTK admin pages.

Finally, as POC, move the UiO-logo top line from VTK to Vortex using this
mechanism.

This commit also relates to: VTK-5350, VTK-5351