[VTK-5431] Introduce 'aclMode' parameter for Repository.create*()
Both createDocument() and createCollection() now take a new argument,
'aclMode', which determines if the resource is created with a
specified ACL, or if the ACL is to be inherired.

If the ACL is explicitly specified, the operations are authorized
using CREATE_WITH_ACL, otherwise the previous authorization behavior
still applies.